Separated boilerplate functions to different files
