+++ src/templates/boilerplate/apiEndpoint.ts
@@ -41,59 +41,3 @@ export class ${pascalCase(endpointClassName)} extends ApiEndpoint {
 }
 `;
 };
-
-export const slashCommandTemplate = (commandName: string): string => {
-    return `
-import { IHttp, IModify, IPersistence, IRead } from '@rocket.chat/apps-engine/definition/accessors';
-import { ISlashCommand, ISlashCommandPreview,
-        ISlashCommandPreviewItem, SlashCommandContext } from '@rocket.chat/apps-engine/definition/slashcommands';
-
-export class ${pascalCase(commandName)} implements ISlashCommand {
-  public command = '';
-  public i18nDescription = '';
-  public i18nParamsExample = '';
-  public permission = '';
-  public providesPreview = false;
-
-  public async executePreviewItem(item: ISlashCommandPreviewItem, context: SlashCommandContext,
-                                  read: IRead, modify: IModify,http: IHttp, persis: IPersistence): Promise<void> {
-      throw new Error('Method not implemented');
-    }
-
-  public async executor(context: SlashCommandContext,read: IRead, modify: IModify,
-                        http: IHttp, persis: IPersistence): Promise<void> {
-      throw new Error('Method not implemented');
-    }
-
-  public async previewer(context: SlashCommandContext, read: IRead, modify: IModify,
-                         http: IHttp, persis: IPersistence): Promise<ISlashCommandPreview> {
-      throw new Error('Method not implemented');
-    }
-}
-`;
-};
-
-export const initialSettingTemplate = (): string => {
-    return `
-import { ISetting, SettingType } from '@rocket.chat/apps-engine/definition/settings';
-export const settings: Array<ISetting> = [
-
-];
-`;
-};
-
-export const appendNewSetting = (data: string): string => {
-    const toWrite = `
-{
-    id: '',
-    type: SettingType.STRING,
-    packageValue: '',
-    required: false,
-    public: false,
-    i18nLabel: '',
-    i18nDescription: '',
-},
-`;
-    const index = data.lastIndexOf('];');
-    return data.slice(0, index) + toWrite + data.slice(index);
-};
