+++ src/templates/boilerplate/index.ts
@@ -1,0 +1,10 @@
+import { apiEndpointTemplate } from './apiEndpoint';
+import { appendNewSetting, initialSettingTemplate } from './setting';
+import { slashCommandTemplate } from './slashCommand';
+
+export {
+    apiEndpointTemplate,
+    appendNewSetting,
+    initialSettingTemplate,
+    slashCommandTemplate,
+};
